IRCTC-Microservices/IRCTC-Microservice-Analyser/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.12.1
annotated-types==0.6.0
anyio==3.6.2
appdirs==1.4.4
APScheduler==3.10.4
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
backcall==0.2.0
beautifulsoup4==4.12.2
blinker==1.4
boto3==1.29.6
botocore==1.32.6
bs4==0.0.1
cachetools==5.3.0
certifi==2022.12.7
cffi==1.16.0
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
comm==0.1.4
command-not-found==0.3
contourpy==1.2.0
croniter==1.4.1
cryptography==3.4.8
cycler==0.12.1
dbus-python==1.2.18
debugpy==1.6.7.post1
decorator==5.1.1
disposable-email-domains==0.0.95
distlib==0.3.8
distro==1.7.0
distro-info===1.1build1
dnspython==2.3.0
et-xmlfile==1.1.0
exceptiongroup==1.1.3
executing==1.2.0
fake-useragent==1.2.1
fastapi==0.92.0
fastapi-cache2==0.2.1
fastapi-utilities==0.1.3.1
filelock==3.12.4
Flask==2.2.3
Flask-Cors==3.0.10
Flask-Login==0.6.2
Flask-Migrate==4.0.5
Flask-Script==2.0.6
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
fonttools==4.47.0
fp==0.2
free-proxy==1.1.1
frozenlist==1.3.3
fsspec==2023.10.0
geoip2==4.6.0
google-api-core==2.15.0
google-api-python-client==2.111.0
google-auth==2.26.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.62.0
greenlet==2.0.2
h11==0.14.0
httplib2==0.21.0
icecream==2.1.3
idna==3.4
importlib-metadata==6.0.0
ipinfo==4.4.2
ipykernel==6.25.1
ipython==8.14.0
itsdangerous==2.1.2
jedi==0.19.0
jeepney==0.7.1
Jinja2==3.1.2
jmespath==1.0.1
jupyter_client==8.3.0
jupyter_core==5.3.1
keyring==23.5.0
kiwisolver==1.4.5
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
Levenshtein==0.21.0
lxml==4.9.3
Mako==1.3.0
MarkupSafe==2.1.2
matplotlib==3.8.2
matplotlib-inline==0.1.6
maxminddb==2.2.0
more-itertools==8.10.0
motor==3.1.1
mpmath==1.3.0
multidict==6.0.4
nest-asyncio==1.5.7
netifaces==0.11.0
networkx==3.2
numpy==1.26.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.52
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.0
openai==0.28.1
openpyxl==3.1.1
openpyxl-image-loader==1.0.5
outcome==1.2.0
packaging==23.1
pandas==2.1.3
parso==0.8.3
pendulum==2.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.1.0
platformdirs==3.10.0
polars==0.20.3
prompt-toolkit==3.0.39
protobuf==4.25.1
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==14.0.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.5.3
pydantic-settings==2.1.0
pydantic_core==2.14.6
pyee==8.2.2
Pygments==2.16.1
PyGObject==3.42.1
PyJWT==2.3.0
pymongo==4.6.0
pymongo-auth-aws==1.1.0
pymongocrypt==1.7.0
pyparsing==3.0.9
pyppeteer==1.0.2
PySocks==1.7.1
python-apt==2.4.0+ubuntu1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.5
pytz==2023.3.post1
pytzdata==2020.1
PyYAML==5.4.1
pyzmq==25.1.1
rapidfuzz==3.2.0
redis==5.0.1
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.7.0
schedule==1.2.1
SecretStorage==3.3.1
selenium==4.11.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
SQLAlchemy==2.0.3
stack-data==0.6.2
starlette==0.25.0
sympy==1.12
systemd-python==234
termcolor==2.2.0
torch==2.1.0
torchaudio==2.1.0
torchvision==0.16.0
tornado==6.3.3
tqdm==4.64.1
traitlets==5.9.0
trio==0.22.2
trio-websocket==0.10.3
triton==2.1.0
typing_extensions==4.9.0
tzdata==2023.3
tzlocal==5.2
ua-parser==0.18.0
ubuntu-advantage-tools==8001
ufw==0.36.1
unattended-upgrades==0.1
undetected-chromedriver==3.5.3
uritemplate==4.1.1
urllib3==1.26.14
user-agents==2.2.0
uvicorn==0.20.0
virtualenv==20.25.0
wadllib==1.3.6
waitress==2.1.2
wcwidth==0.2.6
websockets==10.4
Werkzeug==2.2.3
wsproto==1.2.0
WTForms==3.0.1
xlrd==2.0.1
xlsx2csv==0.8.1
XlsxWriter==3.0.8
xlwt==1.3.0
yarl==1.8.2
zipp==3.13.0
